MVP/pyproject.toml
[project]
name = "mvp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "caldav>=3.1.0",
    "python-dotenv>=1.2.2",
]
